requirements.txt
openai 
requests
matplotlib
Pillow
fpdf2
pyupbit
python-dotenv
selenium
pydantic
telegram
streamlit
schedule
pillow
python-telegram-bot
plotly
pandas
numpy